# [performance] Store temporary hashers in a pool